Create README.md
This project involves scraping data from the Redbus website using Selenium, storing the data in a MySQL database, and visualizing the data using Streamlit. The data scraping is performed in a vs code, the database operations are handled with MySQL, and the visualization is done in Visual Studio Code (VS Code).